condensing information into one section/markdown file
This addresses issue #13 but it may not be what @slhenry and @maryjom
had in mind. I used h3s for the sections for disclaimer, updates, and
terms, but they aren’t styled much different visually than the h2 for
the Introduction. We may not need headings for those subsection, but I
left them in for now.